Introduce system protection(errors)
This commit introduces first mechanics to prevent the user or the hypervisor from self destroying. Namely the following two rules are currently enforced:
 - The root user (0) cannot be deleted
 - The root group (0) cannot be deleted

+++ velocity/db/tables/vDBTGroups.swift
@@ -79,6 +79,10 @@ extension VDB {
 
         /// Removes the group from the database
         func delete() throws {
+            if self.gid == 0 {
+                throw SystemProtectionError(protection: .RootGroupDeletion)
+            }
+
             for media in try self.get_media() {
                 try media.delete()
             }
